requirements.txt
streamlit
PyPDF2
python-docx
pandas
openpyxl
easyocr
sentence-transformers
sqlalchemy
google-generativeai
psycopg2-binary
streamlit-chat
python-dotenv
langchain
langchain-google-genai
faiss-cpu
python-dotenv
pdf2image
PyMuPDF

